Move lastFreeSeatsAnyTraining() to FreeSeats, close to lastFreeSeatsAnyDate()
