+++ site/app/Training/FreeSeats.php
@@ -6,6 +6,7 @@ namespace MichalSpacekCz\Training;
 use DateTime;
 use MichalSpacekCz\Training\Dates\TrainingDate;
 use MichalSpacekCz\Training\Dates\TrainingDateStatus;
+use MichalSpacekCz\Training\Dates\UpcomingTraining;
 use Nette\Database\Row;
 
 class FreeSeats
@@ -34,6 +35,23 @@ class FreeSeats
 		$lastFreeSeats = false;
 		foreach ($dates as $date) {
 			if ($date->isLastFreeSeats()) {
+				$lastFreeSeats = true;
+				break;
+			}
+		}
+		return $lastFreeSeats;
+	}
+
+
+	/**
+	 * @param list<UpcomingTraining>|UpcomingTraining[] $trainings array keys are irrelevant
+	 * @return bool
+	 */
+	public function lastFreeSeatsAnyTraining(array $trainings): bool
+	{
+		$lastFreeSeats = false;
+		foreach ($trainings as $training) {
+			if ($this->lastFreeSeatsAnyDate($training->getDates())) {
 				$lastFreeSeats = true;
 				break;
 			}
